Moved burger to renderhelper
