+++ js/myProfilePage.js
@@ -1,4 +1,4 @@
-import {openMenu, getAllCategories} from './../helperFunctions/renderHelper.js'
+import {openMenu, getAllCategories, burger} from './../helperFunctions/renderHelper.js'
 import {makeRequest,  getUser, verifyAdmin, showCorrectLayout, logOut, printNrOfElements, getAllProducts, getProductFromId} from './../helperFunctions/fetchHelper.js'
 
 
@@ -30,18 +30,6 @@ async function accountCheck() {
 
         return
     }
-}
-
-function burger() {
-
-    const hamburger = document.querySelector(".hamburgerMenu");
-    const menu = document.querySelector(".contactDiv");
-    
-    hamburger.addEventListener("click", () => {
-        hamburger.classList.toggle("active");
-        menu.classList.toggle("active");
-    
-    });
 }
 
 async function getAllLoggedInSubscribers(){
@@ -157,7 +145,6 @@ async function renderOrders(list) {
     let checkAdmin = await verifyAdmin();
 
     if(!checkAdmin) {
-        alert("You are not allowed to proceed")
         return
     }
 
@@ -226,7 +213,6 @@ async function getOrderDetails(id) {
     let checkAdmin = await verifyAdmin();
 
     if(!checkAdmin) {
-        alert("You are not allowed to proceed")
         return
     }
     
@@ -317,6 +303,12 @@ async function getOrderDetails(id) {
 
 async function sendOrder(orderId) {
 
+    let checkAdmin = await verifyAdmin();
+    if(!checkAdmin) {
+        alert("You are not allowed to proceed")
+        return
+    }
+
    let body = new FormData(); 
     body.append("statusId", "SHIP")
     body.append("orderId", orderId)
@@ -367,9 +359,7 @@ document.querySelector(".toggle3").addEventListener("click", () => {
 async function overviewUnitsInStock() {
 
     let checkAdmin = await verifyAdmin();
-
     if(!checkAdmin) {
-        alert("You are not allowed to proceed")
         return
     }
 
@@ -388,7 +378,11 @@ document.querySelector(".getProductButton").addEventListener("click", getUnitsIn
 
 async function getUnitsInStock() {
 
-  /*   let getProduct = document.querySelector(".getProduct") */
+  let checkAdmin = await verifyAdmin();
+  if(!checkAdmin) {
+    alert("You are not allowed to proceed")
+    return
+}
 
     let products = await getAllProducts()
 
@@ -744,7 +738,6 @@ async function getCategoryWithProductId(id) {
     let checkAdmin = await verifyAdmin();
 
     if(!checkAdmin) {
-        alert("You are not allowed to proceed")
         return
     }
 
@@ -960,7 +953,6 @@ async function renderSubscribers() {
     let checkAdmin = await verifyAdmin();
 
     if(!checkAdmin) {
-        alert("You are not allowed to proceed")
         return
     }
